pet clinic adding pet type repository and add dummy data to map

+++ SpringGuru/pet-clinik/pet-clinic-data/src/main/java/com/learn/petclinic/model/PetType.java
@@ -3,7 +3,7 @@ package com.learn.petclinic.model;
 import javax.persistence.Entity;
 
 @Entity
-public class PetType extends BaseModel<Long> {
+public class PetType extends BaseModel<Short> {
 	private String name;
 
 	public PetType(String name) {
